+++ src/dbus/OutputModeService.hpp
@@ -8,7 +8,7 @@ class OutputModeService : public QObject {
     Q_OBJECT
     Q_PROPERTY(int Width READ Width)
     Q_PROPERTY(int Height READ Height)
-    Q_PROPERTY(double RefreshRate READ RefreshRate)
+    Q_PROPERTY(qulonglong RefreshRate READ RefreshRate)
     Q_PROPERTY(bool Preferred READ Preferred)
     Q_PROPERTY(bool Current READ Current)
 public:
@@ -18,7 +18,7 @@ public:
     // Property getters
     int Width() const;
     int Height() const;
-    double RefreshRate() const;
+    qulonglong RefreshRate() const;
     bool Preferred() const;
     bool Current() const;
