+++ src/displays/batch-system/OutputTargetState.hpp
@@ -20,7 +20,7 @@ namespace bd {
         bool isOn() const;
         QSize getDimensions() const;
         QString getMirrorOf() const;
-        double getRefresh() const;
+        qulonglong getRefresh() const;
         QString getRelative() const;
         ConfigurationHorizontalAnchor getHorizontalAnchor() const;
         ConfigurationVerticalAnchor getVerticalAnchor() const;
@@ -28,7 +28,7 @@ namespace bd {
         bool isMirroring() const;
         bool isPrimary() const;
         qreal getScale() const;
-        qint16 getTransform() const;
+        quint8 getTransform() const;
         QSize getResultingDimensions() const;
         uint32_t getAdaptiveSync() const;
 
@@ -36,7 +36,7 @@ namespace bd {
 
         void setOn(bool on);
         void setDimensions(QSize dimensions);
-        void setRefresh(double refresh);
+        void setRefresh(qulonglong refresh);
         void setMirrorOf(const QString& mirrorOf);
         void setRelative(const QString& relative);
         void setHorizontalAnchor(ConfigurationHorizontalAnchor horizontal_anchor);
@@ -44,7 +44,7 @@ namespace bd {
         void setPosition(QPoint position);
         void setPrimary(bool primary);
         void setScale(qreal scale);
-        void setTransform(qint16 transform);
+        void setTransform(quint8 transform);
         void setAdaptiveSync(uint32_t adaptiveSync);
 
         void updateResultingDimensions();
@@ -54,7 +54,7 @@ namespace bd {
         bool m_on;
         QSize m_dimensions;
         QSize m_resulting_dimensions;
-        double m_refresh;
+        qulonglong m_refresh;
         QString m_mirrorOf;
         QString m_relative;
         ConfigurationHorizontalAnchor m_horizontal_anchor;
@@ -62,7 +62,7 @@ namespace bd {
         bool m_primary;
         QPoint m_position;
         qreal m_scale;
-        qint16 m_transform;
+        quint8 m_transform;
         uint32_t m_adaptive_sync;
     };
 }
